feat: add admin controller and views for category model management
