+++ src/main/resources/messages_pt_BR.properties
@@ -4,6 +4,7 @@ nav.dashboard=Dashboard
 nav.entries=Lançamentos
 nav.categories=Categorias
 nav.accounts=Contas
+nav.admin=Administração
 nav.localUsers=Usuários
 
 auth.header.greeting=Olá,
@@ -200,26 +201,6 @@ entries.import.column.category=Categoria
 entries.import.column.amount=Valor
 entries.import.column.notes=Observação
 entries.import.column.externalId=Id Externo
-entries.import.model.title=Modelo de Sugestão de Categoria
-entries.import.model.subtitle=A sugestão usa o último modelo validado e retreina em segundo plano quando necessário.
-entries.import.model.action.retrain=Forçar retreino
-entries.import.model.manualRequested=Retreino solicitado. O último modelo válido continuará sendo usado até a conclusão.
-entries.import.model.status.training=Treinando
-entries.import.model.status.ready=Modelo pronto
-entries.import.model.status.unavailable=Modelo indisponível
-entries.import.model.status.queued=Novo retreino na fila
-entries.import.model.summary.version=Versão
-entries.import.model.summary.modelName=Modelo
-entries.import.model.summary.pipeline=Pipeline
-entries.import.model.summary.examples=Exemplos de treino
-entries.import.model.summary.labels=Categorias
-entries.import.model.summary.vocabulary=Vocabulário
-entries.import.model.summary.trigger=Último gatilho
-entries.import.model.summary.completedAt=Última conclusão
-entries.import.model.trigger.initial=Primeiro treino
-entries.import.model.trigger.csv_import=Importação CSV
-entries.import.model.trigger.manual=Comando manual
-entries.import.model.failure=Última falha registrada: {0}
 entries.statementImport.title=Importar extrato
 entries.statementImport.header.subtitle=Importação de extrato bancário
 entries.statementImport.heading=Importar Extrato
@@ -414,3 +395,27 @@ pagination.nav.last=Última
 theme.toggle.aria=Alternar tema
 theme.toggle.dark=Modo escuro
 theme.toggle.light=Modo claro
+admin.title=Administração
+admin.header.subtitle=Configurações e operações do sistema
+admin.heading=Administração
+admin.subtitle=Centralize configurações operacionais e recursos administrativos do AI$HA.
+admin.models.heading=Administração de Modelos
+admin.models.subtitle=Gerencie os modelos treinados usados pelo sistema. No momento, apenas a sugestão automática de categoria está disponível.
+admin.models.action.retrain=Forçar retreino
+admin.models.manualRequested=Retreino solicitado. O último modelo válido continuará sendo usado até a conclusão.
+admin.models.status.training=Treinando
+admin.models.status.ready=Modelo pronto
+admin.models.status.unavailable=Modelo indisponível
+admin.models.status.queued=Novo retreino na fila
+admin.models.summary.version=Versão
+admin.models.summary.modelName=Modelo
+admin.models.summary.pipeline=Pipeline
+admin.models.summary.examples=Exemplos de treino
+admin.models.summary.labels=Categorias
+admin.models.summary.vocabulary=Vocabulário
+admin.models.summary.trigger=Último gatilho
+admin.models.summary.completedAt=Última conclusão
+admin.models.trigger.initial=Primeiro treino
+admin.models.trigger.csv_import=Importação CSV
+admin.models.trigger.manual=Comando manual
+admin.models.failure=Última falha registrada: {0}
