+++ src/main/resources/templates/fragments/header.html
@@ -38,6 +38,10 @@
                         <i data-lucide="briefcase" aria-hidden="true"></i>
                         <span th:text="#{nav.accounts}"></span>
                     </a>
+                    <a th:href="@{/admin}">
+                        <i data-lucide="shield-check" aria-hidden="true"></i>
+                        <span th:text="#{nav.admin}"></span>
+                    </a>
                     <a th:href="@{/local-users}">
                         <i data-lucide="users" aria-hidden="true"></i>
                         <span th:text="#{nav.localUsers}"></span>
