+++ src/test/java/dev/ccosta/aisha/security/SecurityIntegrationTest.java
@@ -1,8 +1,10 @@
 package dev.ccosta.aisha.security;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.Matchers.containsString;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
@@ -43,6 +45,15 @@ class SecurityIntegrationTest {
     }
 
     @Test
+    void shouldRenderAdminPageWhenAuthenticated() throws Exception {
+        HttpSession session = loginAndGetSession();
+
+        mockMvc.perform(get("/admin").session((MockHttpSession) session))
+            .andExpect(status().isOk())
+            .andExpect(content().string(containsString("Administração de Modelos")));
+    }
+
+    @Test
     void shouldRejectProtectedPostWithoutCsrf() throws Exception {
         HttpSession session = loginAndGetSession();
 
@@ -64,6 +75,17 @@ class SecurityIntegrationTest {
                 .param("redirectTo", "/dashboard"))
             .andExpect(status().isFound())
             .andExpect(redirectedUrl("/dashboard"));
+    }
+
+    @Test
+    void shouldAllowRetrainRequestFromAdminPageWithCsrf() throws Exception {
+        HttpSession session = loginAndGetSession();
+
+        mockMvc.perform(post("/admin/category-model/retrain")
+                .session((MockHttpSession) session)
+                .with(csrf()))
+            .andExpect(status().isOk())
+            .andExpect(content().string(containsString("Retreino solicitado")));
     }
 
     @Test
